test: rename app-layout first-paint tests to content-offset
